% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMexpression}
\alias{IMexpression}
\title{Infectious mononucleosis transcriptome}
\format{A matrix with 100 rows and 89 variables}
\usage{
data(IMexpression)
}
\description{
This is compilied numeric matrix of raw microarray intensities data. Each of 100 rows corresponds to a
probe (gene, transcript) and each of 89 column correspondes to a specimen (patient). Specimen is total mRNA
samples from human peripheral blood leukocytes' taken from healthy children and children with
infectious mononucleosis of different ethiology.
Specimen features are in \code{IMspecimen} data.
}
\details{
\itemize{
  \item rownames. Probes IDs. For most transcripts there are several probes.
  \item colnames. Specimen IDs.
  }
}
\author{
Nikolai A. Sakharnov, Dmitry I. Knyazev, Oleg V. Utkin
}
\keyword{datasets}
