% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMspecimen}
\alias{IMspecimen}
\title{Specimen features}
\format{A data frame with 89 rows and 2 variables.}
\usage{
data(IMspecimen)
}
\description{
A dataset containing information about specimens from \code{IMexpression} data: IDs and diagnoses.
}
\details{
\itemize{
  \item ID. Specimen ID, factor variable with 89 levels.
  \item diagnosis. Specimen diagnosis:
  "ebv" (children with acute EBV mononucleosis),
  "hhv6" (children with acute HHV6 mononucleosis),
  "norm" (healthy children). Factor variable with 3 levels.
}
}
\author{
Nikolai A. Sakharnov, Dmitry I. Knyazev, Oleg V. Utkin
}
\keyword{datasets}
